Move test around so mutant will be able to find them when needed
